changes
After testing the API, I decided to delete some files because they didn't fit well with the logic of the project. Similarly, I removed the reservefood feature because it didn't work as I had hoped. Then, I made some improvements by linking tables together. For instance, I made it possible to create invoices even without existing orders.

Regarding payments, I realized that it duplicated information already present in the invoice, so I removed that feature. Additionally, I added a staffaccount feature and included information about allergies, which I hadn't thought of earlier. Lastly, I included some files in the initialization process that I had forgotten to add before.

+++ api/deleteinvoice.php
@@ -1,16 +1,30 @@
 <?php
+
+header('Access-Control-Allow-Origin: *');
+header('Content-Type: application/json');
+
+header('Access-Control-Allow-Methods: DELETE');
+header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization, X-Requested-With');
+
 include_once('../core/initialize.php');
 
-$invoice = new Invoice($db);
-$data = json_decode(file_get_contents("php://input"));
-
-// Set invoice ID to be deleted
-$invoice->invoiceId = $data->invoiceId;
-
-// Delete the invoice
-if ($invoice->delete()) {
-    echo json_encode(array('message' => 'Invoice deleted'));
-} else {
-    echo json_encode(array('message' => 'Failed to delete invoice'));
+if (!isset($_GET['invoiceId'])) {
+    echo json_encode(array('message' => 'Failed to delete invoice. Invoice ID is missing.'));
+    exit;
 }
+
+$invoiceId = $_GET['invoiceId'];
+
+$invoice = new Invoice($db);
+
+// Assign the invoiceId to the Invoice object
+$invoice->invoiceId = $invoiceId;
+
+// Attempt to delete the invoice
+if ($invoice->delete()) {
+    echo json_encode(array('message' => 'Invoice deleted successfully.'));
+} else {
+    echo json_encode(array('message' => 'Failed to delete invoice.'));
+}
+
 ?>
